requirements.txt
asgiref==3.5.2
dj-rest-auth==2.2.5
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Markdown==3.4.1
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
